esmi_ib_library: simplify getting sibling core status
When SMT is enabled, Linux enumerates all the sibling threads of a core
as CPUs. Since, the power management information is available at
core level, the linux drivers expose the information for the main
core id. If the user requests the information for the sibling thread,
identify the main core id in the library.

Signed-off-by: Muralidhara M K <muralidhara.mk@amd.com>