e_smi_library: Move the PKG_BOOSTLIMIT_TYPE entry
Package boost limit is not a hsmp message and in proto ver 5, 13h is
not a reserved message, hence need to move this.

Once sysfs support is removed, we can safely remove this also.
Since sysfs had this boost limit file in common across the package
this enum entry was added in library. It internally uses
SOCKET_BOOST_LIMIT hsmp message to write to the available sockets.

Signed-off-by: Suma Hegde <suma.hegde@amd.com>