lib: pick energy driver based on rapl register size
In AMD EPYC family, some of the initial processor models support
only 32 bit RAPL registers and energy accumulation
was needed for these models because of wrap around.
So in that case, we use "amd_energy" driver which takes care of this
problem.

For other EPYC models RAPL registers are already 64 bit wide and we dont
need accumulation. Hence we use "msr-safe" driver available open-source

In this commit, just restructuring the code to reflect the above
changes.