lib: Add amd_hsmp.h as part of esmi repo
amd_hsmp.h is needed for esmi compilation. It should come as part of linux headers.
But if there is a mismatch in the version of header required by esmi and which is
installed as part of linux, then error will be thrown during esmi compilation.
Hence add this file locally.